rename create.py to writer for name consistency
