+++ src/somesy/codemeta/__init__.py
@@ -1,4 +1,4 @@
 """Integration with codemetapy (to re-generate codemeta as part of somesy sync)."""
-from .create import Codemeta
+from .writer import Codemeta
 
 __all__ = ["Codemeta"]
