+++ src/somesy/codemeta/writer.py
